Fix differenceInXX functions, migrate differenceInMinutes to TS
